Factor out helper for getting C runtime linkage
